Simplify structures
